Added SlTrace/TraceControlWindow Button support
SlTrace.traceButton(flag, function) facilitates placing a button  which will call the function to allow independence in invoking debugging operations such as a reset of state at operator's wish.  It came in handy to reset the undo/redo player control info stacks in the dots project when I was debugging.